+++ CHANGELOG.rst
@@ -17,6 +17,16 @@ Unreleased
 ----------
 * nothing unreleased
 
+[8.13.0] - 2026-09-14
+----------------------
+* Changed ``SupportEnterpriseEnrollmentDataInjector`` to accept/return the full enrollments list
+  and attach enterprise course enrollment data to each matching enrollment dict in place, instead
+  of returning a separate dict keyed by course_id that the platform had to zip back together itself
+
+[8.12.0] - 2026-09-08
+----------------------
+* feat: add SupportEnterpriseEnrollmentDataInjector pipeline step (ENT-11574)
+
 [8.11.0] - 2026-09-10
 ----------------------
 * fix: change SupportContactEnterpriseTagStep to accept/return full context (ENT-11574)